+++ packages/dagonizer-embedder-mistral/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @studnicky/dagonizer-embedder-mistral
 
+## 1.0.1
+
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies [fdaa32a]
+  - @studnicky/dagonizer@1.0.0
+
+## 0.30.1
+
+## 0.30.0
+
+### Patch Changes
+
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+  - @studnicky/dagonizer@1.0.0
+
 ## 0.30.0
 
 ## 0.29.1