+++ packages/dagonizer-store-opfs/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @studnicky/dagonizer-store-opfs
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [8a57ce4]
+  - @studnicky/dagonizer@1.0.1
+
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies [fdaa32a]
+  - @studnicky/dagonizer@1.0.0
+
+## 0.30.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @studnicky/dagonizer@0.30.1
+
+## 0.30.0
+
+### Patch Changes
+
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+- Updated dependencies [4639f9b]
+  - @studnicky/dagonizer@1.0.0
+
 ## 0.30.0
 
 ### Patch Changes